requirements.txt
crewai[google-genai]
crewai-tools
python-dotenv
pyyaml
